+++ src/canister/src/storage/group_events_storage.rs
@@ -1,4 +1,4 @@
-use super::storage_api::{StorageMethods, CLEAR_MEMORY_ID, GROUP_EVENTS, MEMORY_MANAGER};
+use super::storage_api::{StorageMethods, GROUP_EVENTS, GROUP_EVENTS_MEMORY_ID, MEMORY_MANAGER};
 use canister_types::models::{api_error::ApiError, event_collection::EventCollection};
 use ic_stable_structures::StableBTreeMap;
 
@@ -137,7 +137,7 @@ impl StorageMethods<u64, EventCollection> for GroupEventsStore {
     fn clear() -> () {
         GROUP_EVENTS.with(|n| {
             n.replace(StableBTreeMap::new(
-                MEMORY_MANAGER.with(|m| m.borrow().get(CLEAR_MEMORY_ID)),
+                MEMORY_MANAGER.with(|m| m.borrow().get(GROUP_EVENTS_MEMORY_ID)),
             ))
         });
     }